re-factor global options and statistics
Program options set by command line or configuration file
are shared between all modules, as are global statistics that
are collected for runtime measures.

This patch collects all global variables into related objects,
and exposing them via public pointers.

This is a preparational work to modularize the handling of
configuration and runtime statistics.