working version of stats, ztex clock changes, pico support
The problem was the order of include files.  I was including the pico
driver header files before the header file with my structure.  The main
cgminer program didn't include (obviously) the pico driver header files.
I had to comment out a 'pragma align = reset' in the pico driver because
gcc didn't understand it.  This was probably the root of the issue.