Add --cert option to allow self-certified https: JSON-RPC servers
When using distributed miners to a central coin server, the RPC username
and password travel unencrypted over the net, allowing any snooper to empty
your wallet.  To defeat this, we would use https: but OpenSSL fails if the
certificate is self-signed, which is the normal case for weebs like me.

This patch adds a --cert option allowing users to supply the RPC server's
self-signed certificate.